backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
twilio
pyjwt
passlib[bcrypt]
redis
requests